# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LocalAI Dev Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/conversations:
    get:
      operationId: listConversations
      tags: [ai]
      summary: List AI conversations
      parameters:
        - name: workspaceId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [ai]
      summary: Start a new AI conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /ai/conversations/{id}:
    get:
      operationId: getConversation
      tags: [ai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationWithMessages"
    delete:
      operationId: deleteConversation
      tags: [ai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /ai/conversations/{id}/messages:
    post:
      operationId: sendMessage
      tags: [ai]
      summary: Send a message and get AI response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "200":
          description: AI response message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
  /ai/models/{name}/pull:
    post:
      operationId: pullAiModel
      tags: [models]
      summary: Pull/download a model
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pull status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelPullStatus"
  /ai/models/{name}:
    delete:
      operationId: deleteAiModel
      tags: [models]
      summary: Delete a local model
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Deployments ───────────────────────────────────────
  /deployments:
    get:
      operationId: listDeployments
      tags: [deployments]
      summary: List deployments
      parameters:
        - name: workspaceId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of deployments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deployment"
    post:
      operationId: createDeployment
      tags: [deployments]
      summary: Trigger a new deployment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeploymentInput"
      responses:
        "201":
          description: Created deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /deployments/{id}:
    get:
      operationId: getDeployment
      tags: [deployments]
      summary: Get deployment details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /deployments/{id}/rollback:
    post:
      operationId: rollbackDeployment
      tags: [deployments]
      summary: Rollback to this deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: New rollback deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
...
  /containers:
    get:
      operationId: listContainers
      tags: [containers]
      summary: List running containers
      responses:
        "200":
          description: List of containers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Container"
    post:
      operationId: createContainer
      tags: [containers]
      summary: Create a new container
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContainerInput"
      responses:
        "201":
          description: Created container
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Container"
  /containers/{id}:
    get:
      operationId: getContainer
      tags: [containers]
      summary: Get container details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Container
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Container"
    delete:
      operationId: stopContainer
      tags: [containers]
      summary: Stop and remove a container
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Stopped
  /containers/{id}/logs:
    get:
      operationId: getContainerLogs
      tags: [containers]
      summary: Get container logs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Container log lines
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContainerLogs"

  # ── Metrics ───────────────────────────────────────────
...
components:
  schemas:
...
    Conversation:
      type: object
      required: [id, title, model, messageCount, createdAt, updatedAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: ["integer", "null"]
        title:
          type: string
        model:
          type: string
        messageCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    ConversationInput:
      type: object
      required: [title, model]
      properties:
        title:
          type: string
        model:
          type: string
        workspaceId:
          type: integer
    ConversationWithMessages:
      type: object
      required: [id, title, model, messages, createdAt, updatedAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: ["integer", "null"]
        title:
          type: string
        model:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        createdAt:
          type: string
        updatedAt:
          type: string
    Message:
      type: object
      required: [id, conversationId, role, content, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        tokensUsed:
          type: ["integer", "null"]
        createdAt:
          type: string
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        model:
          type: string

    # ── AI Models ─────────────────────────────────────
...
    ModelPullStatus:
      type: object
      required: [name, status, progress]
      properties:
        name:
          type: string
        status:
          type: string
        progress:
          type: integer

    # ── Deployments ───────────────────────────────────
    Deployment:
      type: object
      required: [id, workspaceId, status, environment, createdAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        workspaceName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, building, deploying, success, failed, rolled_back]
        environment:
          type: string
          enum: [development, staging, production]
        url:
          type: ["string", "null"]
        commitSha:
          type: ["string", "null"]
        commitMessage:
          type: ["string", "null"]
        buildDuration:
          type: ["integer", "null"]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
        finishedAt:
          type: ["string", "null"]
    DeploymentInput:
      type: object
      required: [workspaceId, environment]
      properties:
        workspaceId:
          type: integer
        environment:
          type: string
          enum: [development, staging, production]
        commitSha:
          type: string
        commitMessage:
          type: string

    # ── Containers ────────────────────────────────────
    Container:
      type: object
      required: [id, workspaceId, name, image, status, createdAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        name:
          type: string
        image:
          type: string
        status:
          type: string
          enum: [running, stopped, paused, exited, creating]
        port:
          type: ["integer", "null"]
        cpuPercent:
          type: ["number", "null"]
        memoryMb:
          type: ["integer", "null"]
        createdAt:
          type: string
        startedAt:
          type: ["string", "null"]
    ContainerInput:
      type: object
      required: [workspaceId, name, image]
      properties:
        workspaceId:
          type: integer
        name:
          type: string
        image:
          type: string
        port:
          type: integer
        envVars:
          type: object
          additionalProperties:
            type: string
    ContainerLogs:
      type: object
      required: [containerId, lines]
      properties:
        containerId:
          type: integer
        lines:
          type: array
          items:
            type: object
            required: [timestamp, stream, message]
            properties:
              timestamp:
                type: string
              stream:
                type: string
                enum: [stdout, stderr]
              message:
                type: string

    # ── Metrics ───────────────────────────────────────
...