# lib/api-zod/src/generated/types/container.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LocalAI Dev Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ContainerStatus } from './containerStatus';

export interface Container {
  id: number;
  workspaceId: number;
  name: string;
  image: string;
  status: ContainerStatus;
  /** @nullable */
  port?: number | null;
  /** @nullable */
  cpuPercent?: number | null;
  /** @nullable */
  memoryMb?: number | null;
  createdAt: string;
  /** @nullable */
  startedAt?: string | null;
}